Add "add-many" test and refactor parsing
We're going to use the antlr labels.  There are going to so many
EnterXXX() and ExitXXX() functions.  I removed the base parser and
implemented all the required methods because I couldn't get one of the
methods to be called, after repeatedly checking the name and
arguments.  The parser is now broken up into multiple files.